backend/requirements.txt
# FleetFlow Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.0
supabase>=2.3.0
python-dotenv>=1.0.0
